content-security-policy/to-port-from-blink/1.1/reflected-xss-and-xss-protection-invalid-block-expected.txt

CONSOLE ERROR: The 'reflected-xss' Content Security Policy directive has the invalid value "invalid". Valid values are "allow", "filter", and "block".
CONSOLE ERROR: line 4: The XSS Auditor blocked access to 'http://127.0.0.1:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Cscript%3Ealert(String.fromCharCode(0x58,0x53,0x53))%3C/script%3E&csp=invalid&enable-full-block=1' because the source code of a script was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
ALERT: Loaded http://127.0.0.1:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Cscript%3Ealert(String.fromCharCode(0x58,0x53,0x53))%3C/script%3E&csp=invalid&enable-full-block=1 into the IFrame.
Testing behavior when "reflected-xss" is set to invalid, and "X-XSS-Protection" is set to block.

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
